+++ sdk/resourcemanager/astro/armastro/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 1.0.0 (2024-06-22)
+### Breaking Changes
+
+- Type of `OrganizationResourceUpdate.Identity` has been changed from `*ManagedServiceIdentity` to `*AzureResourceManagerLegacyManagedServiceIdentityV4Update`
+
+### Features Added
+
+- New struct `AzureResourceManagerLegacyManagedServiceIdentityV4Update`
+
+
 ## 0.1.0 (2024-02-23)
 
 The package of `github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/astro/armastro` is using our [next generation design principles](https://azure.github.io/azure-sdk/general_introduction.html).